Skeletal Unicode loading, go generate
Using crude shell to make the file which embeds the raw Unicode data in
a `.go` file.  For now, storing the data as one big contiguous `[]byte`
blob.  It's largish, but still less than 2MB and would be loaded and
parsed in a scripting language anyway, so overhead minimal.

Planning to be inefficient with parsing/referencing for the time being.
If I don't mess up the interfaces too much, should be able to optimize
later.