Ferret version update (fixed reported bug)
Hint at invoking to use substring search in the top-level README.

Ferret upstream fixed my reported bug and in so doing also made my
workaround no longer work.  (Glad that I documented that this workaround
wasn't guaranteed by documented contract.)

So, pull in the fixed version and update Deps to lock to it.