Unbreak tests re resultset & output
resultset had switched to the idea of selectable output destinations and
setting the destinations if nil to be stdout/stderr.  This worked,
except that the fragile test messing with stdout/stderr no longer
worked.

Export the stream names, and actually use the new stream support for the
test.  Rips out a bunch of dependencies and code, test is much simpler,
proving out the change.

Memo to self: run tests before push.  And before signing a tag.