Show block information for characters
Parse the block information into an array of <300 items, and search it
linearly each time.  Not algorithmically clean, but it's fast enough for
data this small.

Made the render-one-char-one-attribute-to-string function be an exposed
method rather than an internal function, so that we can easily lookup
blocks, etc.  Might rethink this, might be best to leave callers just
using LookupBlock instead.