Add the browse subcommand
Highlights that we don't handle right-to-left well.

The PRINT_RUNE_ISOLATED code isn't working, so is unused, but left in to
try to figure out what should be done.

The default behaviour, to browse everything, exposes a pathological
behaviour of termtables (it takes 3 minutes to render, for me).

Exposes that vim digraph lookups are broken for "more than single
character shown as char".