code: lookup handling hex-encoded UTF-8
Add `-H` / `--utf8hex` flag: take UTF-8 Hex-encoded codepoint

Examples:

    character code -vH caf%C3%A9
    character code -vH %CE%BA%CE%B1%CF%86%CE%AD
    character code -cvH CEBACEB1CF86CEAD
    character name -v καφέ

If a given arg contains `%` then it's assumed to be URL-style encoded
and we take all non-%HH characters as literals, effectively acting like
`character name` for those characters, but mixing in % decoding too.

If `%` is not present, then treat as a pure hex string, with runes
encoded as UTF-8 and then in hex.

Open question: should this really be in the `name` command, instead of
`code`?  I thought of it as "taking the number but decoding it first".
Might be wrong.