Add second table provider: olekukonko/tablewriter
<https://github.com/olekukonko/tablewriter> provides a Golang table
driver somewhat similar to early stages of
<https://github.com/apcera/termtables>, before I gave it UTF-8 support.

Split apart the tables setup, use build tags to select a backend, define
one as the default in build tags (for `go build`), and explicitly set
one as the default in the GNUmakefile.  Pass through appropriate build
tags on the command-line, taking advantage of loose whitespace-splitting
by `go build` of the `-tags` value.