Support for other encodings
Add hex-decode to the `name` sub-command too.
Add an encoding flag, taking any of a list of supported names.
Add a list-encodings, which has the canonical encoding names.

We don't list all variants, but do support them.  Would rather have a
couple of aliases be supported, undocumented, than withdraw them later.
They have to be explicitly requested.

In combination:

    character name -e Shift-JIS -Hv =01=02=03

but with real hex input for Shift-JIS