Add `Of` column for decomposed; fix tests
To make it clearer where characters come from, now that they might not
map exactly to the cmdline, add an "Of" column to the end of the two
main tables and use it to display the original input character where it
differs.

Tests broken by language change around "." as import source?