New 'known' subcommand; move two lists therein
To add a `puny` command, I want to reuse the encodings support because
needing non-UTF8 encoding for input to the punycode encoding is more
likely.  This commit is preliminary to that; no `puny` yet, but we do
add `known`.

Move `table-list-styles` sub-command to `deprecated/`, hidden and now
implemented as an alias for the real code.

Add `known`: default behavior is to show usage, but using a flag to ask
for something will inhibit that.  List table styles, list encodings.