Performance compromise: pre-gen some Unicode data
Switch the Unicode pre-generated data to be only linear lists, not the
maps too.  Populate the maps at runtime.

Shrinks compilation time from 40s down to 2s.  Not a corresponding
major performance hit in runtime.

Speculation: compilation is only extending the size of maps a little at
a time each time, so when there are 31618 keys in each of a couple of
maps, things Behave Poorly