Travis: stop building on older Golang: bad compiler
The new unicode/ sub-package's pre-generated data structures trigger
really bad performance in Go 1.7.x and older; about as bad as the
current code did in 1.8.1 when I had the dual 30k element maps in there.
Am assuming this is some bad performance with adding to the lists.

So just stop testing on older Golang, since it does build, but slowly
enough that Travis either times out and kills it or is setting resource
limits such that we die; I can't tell for sure.