go module; unbreak test "broken" by upstream
Go module support.

One of the many updates to the excellent `github.com/mattn/go-runewidth`
was commit `afa37cd0`, "use unicode9 widths", from November 2016.  That
reclassified one of the test characters as an emoji and changed the
width.

_This_ tool is maintained, but stable and I hadn't had cause to invoke
the test suite recently on a new box.  Without go mod, I probably kept
using the old version of go-runewidth.  So, win.