Simplify with use of `go list`
To get the import path of the current repo, we can just use `go list`
which is equivalent to `go list -f '{{.ImportPath}}'`

Nuke REPO_PATH, CANONICAL_CHARACTER_REPO, simplify.

This might introduce infelicities in the handling of the pre-module
builds.  Not too much to be done about that today.

Bonus: if we're ever tagged `v2.x.y` then our import paths will still be
correct.