Rune width fixes again
The root cause of some display issues I saw turned out to be that I'd
missed that `go.mod` had latched onto an ancient version of
go-runewidth; I fixed that and all my woes were resolved.

For characters in the various graphical ranges, override to 2 chars
wide, not 1.  This might invert the systems where the results are
"correct", time will tell.

The not-yet-working PRINT_RUNE_ISOLATED gets put inside width-specifying
units, so that at least that does not regress when isolating directions.

Add `renderWidth` to the JSON output.

Drop the `Of` column from the normal output view, leave it in the Net
view; technically it's not net, and it should always be present, but
really this requires getting working auto-suppression in tabular output.

Adjusted the `--internal-debug` table columns.