Add --shell-parse-argv and use for WASM web-page
The WASM index page was splitting params on space, which worked as a
quick hack but left us with no way to ask for something like:

    browse -b 'Miscellaneous Symbols and Arrows'

Rather than try to figure out bundling for shell parsers for JS, keep
our main logic in one language and the JS as a light shim.  Add a new
global option `--shell-parse-argv`, designed to be the second-last item
in argv, with a single string in the last item in argv, the option
value.  When encountered, split per shell rules and replace the argv for
cobra, triggering a re-parse.

The method I found to get Cobra to re-parse, I'm going to describe as
"future idiomatic".  That sounds better than "gross hack", right?