Add --left/--right to render flags
The actual rendering of <http://unicode.org/reports/tr51/#Direction> sequences
leaves a little to be desired at present, in various apps, but we can now emit
it!

But it messes up our display when we do so, since the tool can't interrogate
the output sequence down the stack of processes to a tty or other sink to ask
about support for certain sequences, so we just assume that if --left/--right
is asked for, then it will render correctly.